refactor: structure prepare
