+++ AI_chatbot/consts/dl_model_training.py
@@ -1,0 +1,12 @@
+DEFAULT_BATCH_SIZE = 8
+DEFAULT_NUM_WORKES = 0
+DEFAULT_OPTIMIZER = "adam"
+DEFAULT_LOSS = "crossentropyloss"
+DEFAULT_LEARNING_RATE = 3e-4
+DEFAULT_HIDDEN_SIZE = 64
+DEFAULT_MAX_EPOCHS = 500
+DEFAULT_INPUT_SIZE = 5
+DEFAULT_OUTPUT_SIZE = 10
+
+RANDOM_STATE = 42
+VAL_DATASET_SIZE = 0.2
