+++ AI_chatbot/consts/logging.py
@@ -1,0 +1,1 @@
+FORMAT = "%(asctime)s:%(pathname)s:%(levelname)s:%(lineno)d:%(message)s"
